+++ lib/src/ast/css/media_query.dart
@@ -2,6 +2,7 @@
 // MIT-style license that can be found in the LICENSE file or at
 // https://opensource.org/licenses/MIT.
 
+import '../../interpolation_map.dart';
 import '../../logger.dart';
 import '../../parse/media_query.dart';
 import '../../utils.dart';
@@ -43,8 +44,10 @@ class CssMediaQuery {
   ///
   /// Throws a [SassFormatException] if parsing fails.
   static List<CssMediaQuery> parseList(String contents,
-          {Object? url, Logger? logger}) =>
-      MediaQueryParser(contents, url: url, logger: logger).parse();
+          {Object? url, Logger? logger, InterpolationMap? interpolationMap}) =>
+      MediaQueryParser(contents,
+              url: url, logger: logger, interpolationMap: interpolationMap)
+          .parse();
 
   /// Creates a media query specifies a type and, optionally, conditions.
   ///