+++ lib/src/ast/selector/pseudo.dart
@@ -5,6 +5,7 @@
 import 'package:charcode/charcode.dart';
 import 'package:collection/collection.dart';
 import 'package:meta/meta.dart';
+import 'package:source_span/source_span.dart';
 
 import '../../utils.dart';
 import '../../util/nullable.dart';
@@ -104,11 +105,12 @@ class PseudoSelector extends SimpleSelector {
     }
   }();
 
-  PseudoSelector(this.name,
+  PseudoSelector(this.name, FileSpan span,
       {bool element = false, this.argument, this.selector})
       : isClass = !element && !_isFakePseudoElement(name),
         isSyntacticClass = !element,
-        normalizedName = unvendor(name);
+        normalizedName = unvendor(name),
+        super(span);
 
   /// Returns whether [name] is the name of a pseudo-element that can be written
   /// with pseudo-class syntax (`:before`, `:after`, `:first-line`, or
@@ -135,14 +137,15 @@ class PseudoSelector extends SimpleSelector {
 
   /// Returns a new [PseudoSelector] based on this, but with the selector
   /// replaced with [selector].
-  PseudoSelector withSelector(SelectorList selector) => PseudoSelector(name,
-      element: isElement, argument: argument, selector: selector);
+  PseudoSelector withSelector(SelectorList selector) =>
+      PseudoSelector(name, span,
+          element: isElement, argument: argument, selector: selector);
 
   /// @nodoc
   @internal
   PseudoSelector addSuffix(String suffix) {
     if (argument != null || selector != null) super.addSuffix(suffix);
-    return PseudoSelector(name + suffix, element: isElement);
+    return PseudoSelector(name + suffix, span, element: isElement);
   }
 
   /// @nodoc
@@ -200,7 +203,8 @@ class PseudoSelector extends SimpleSelector {
 
     // Fall back to the logic defined in functions.dart, which knows how to
     // compare selector pseudoclasses against raw selectors.
-    return CompoundSelector([this]).isSuperselector(CompoundSelector([other]));
+    return CompoundSelector([this], span)
+        .isSuperselector(CompoundSelector([other], span));
   }
 
   T accept<T>(SelectorVisitor<T> visitor) => visitor.visitPseudoSelector(this);