+++ lib/src/extend/empty_extension_store.dart
@@ -3,16 +3,17 @@
 // https://opensource.org/licenses/MIT.
 
 import 'package:collection/collection.dart';
-import 'package:source_span/source_span.dart';
 import 'package:tuple/tuple.dart';
 
 import '../ast/css.dart';
-import '../ast/css/modifiable.dart';
 import '../ast/selector.dart';
 import '../ast/sass.dart';
+import '../util/box.dart';
 import 'extension_store.dart';
 import 'extension.dart';
 
+/// An [ExtensionStore] that contains no extensions and can have no extensions
+/// added.
 class EmptyExtensionStore implements ExtensionStore {
   bool get isEmpty => true;
 
@@ -24,15 +25,14 @@ class EmptyExtensionStore implements ExtensionStore {
           bool callback(SimpleSelector target)) =>
       const [];
 
-  ModifiableCssValue<SelectorList> addSelector(
-      SelectorList selector, FileSpan span,
+  Box<SelectorList> addSelector(SelectorList selector,
       [List<CssMediaQuery>? mediaContext]) {
     throw UnsupportedError(
         "addSelector() can't be called for a const ExtensionStore.");
   }
 
   void addExtension(
-      CssValue<SelectorList> extender, SimpleSelector target, ExtendRule extend,
+      SelectorList extender, SimpleSelector target, ExtendRule extend,
       [List<CssMediaQuery>? mediaContext]) {
     throw UnsupportedError(
         "addExtension() can't be called for a const ExtensionStore.");
@@ -43,7 +43,6 @@ class EmptyExtensionStore implements ExtensionStore {
         "addExtensions() can't be called for a const ExtensionStore.");
   }
 
-  Tuple2<ExtensionStore,
-          Map<CssValue<SelectorList>, ModifiableCssValue<SelectorList>>>
-      clone() => const Tuple2(EmptyExtensionStore(), {});
+  Tuple2<ExtensionStore, Map<SelectorList, Box<SelectorList>>> clone() =>
+      const Tuple2(EmptyExtensionStore(), {});
 }