+++ pkg/sass_api/lib/sass_api.dart
@@ -17,6 +17,7 @@ export 'package:sass/src/ast/selector.dart';
 export 'package:sass/src/async_import_cache.dart';
 export 'package:sass/src/exception.dart' show SassFormatException;
 export 'package:sass/src/import_cache.dart';
+export 'package:sass/src/interpolation_map.dart';
 export 'package:sass/src/value.dart' hide ColorFormat, SpanColorFormat;
 export 'package:sass/src/visitor/find_dependencies.dart';
 export 'package:sass/src/visitor/interface/expression.dart';
@@ -26,6 +27,7 @@ export 'package:sass/src/visitor/recursive_ast.dart';
 export 'package:sass/src/visitor/recursive_selector.dart';
 export 'package:sass/src/visitor/recursive_statement.dart';
 export 'package:sass/src/visitor/replace_expression.dart';
+export 'package:sass/src/visitor/selector_search.dart';
 export 'package:sass/src/visitor/statement_search.dart';
 
 /// Parses [text] as a CSS identifier and returns the result.